chore: Update npm dependencies to latest stable versions and add react-helmet
This commit updates the npm dependencies in the package.json and package-lock.json files to their latest stable versions. It also adds the "react-helmet" package as a new dependency. This update ensures that the project is using the most up-to-date versions of the dependencies and includes the necessary package for using the "react-helmet" library.